# 🏥 Optimize Health Check Logging - Reduce Log Noise by 50%