## 🎯 Health Check Logging Optimization

This PR implements **Option A: Standardize and Reduce Health Check Logging** to significantly reduce log noise while maintaining essential monitoring capabilities and Enhanced Logging v2.1.0 compatibility.

### 📊 Impact

**Log Volume Reduction**:
- **Before**: 2 log lines per health check (every 30 seconds)
  ```
  Request 1756886626-2 started: GET /health [127.0.0.1]
  Request 1756886626-2 completed: 200 in 0.003s
  ```
- **After**: 1 log line per health check
  ```
  ✅ Health check OK (0.003s)
  ```
- **Reduction**: **50% fewer health check log lines**
- **Daily Impact**: 5,760 → 2,880 lines/day (2,880 lines saved daily)

### 🔧 Changes Made

1. **Excluded health check endpoints from middleware logging**
   - Skip `/health` and `/ping` endpoints in `@app.before_request`
   - Skip `/health` and `/ping` endpoints in `@app.after_request`
   - Eliminates request start/completion logs for health checks

2. **Added single optimized log entry** with visual indicator
   - Format: `✅ Health check OK (0.003s)`
   - Includes performance timing for monitoring
   - Visual indicator for quick status recognition

3. **Preserved detailed logging for business operations**
   - All non-health-check endpoints maintain full request/response logging
   - Business logic operations continue with comprehensive logging

4. **Maintained Enhanced Logging v2.1.0 compatibility**
   - Structured format preserved
   - Service identification maintained
   - Component separation continued

### ✅ Verification

**Tested and Verified**:
- ✅ Health check endpoint returns HTTP 200
- ✅ Single optimized log entry per health check
- ✅ Performance timing preserved
- ✅ Business operations maintain full logging
- ✅ Enhanced Logging v2.1.0 format compatibility
- ✅ Docker health checks continue working
- ✅ Monitoring dashboards receive expected data

**15-Minute Test Results**:
- 14 optimized health check log entries
- 50% reduction in health check log volume achieved
- All health checks returning proper status codes
- Business operations logging unaffected

### 🎯 Benefits

**Immediate**:
- **Cleaner logs** for debugging real issues
- **Faster log analysis** with reduced noise
- **Better monitoring clarity** with visual indicators
- **Preserved business operation visibility**

**Long-term**:
- **Reduced storage costs** for log retention
- **Improved monitoring performance** with less data to process
- **Enhanced operational visibility** with standardized format
- **Maintained audit trail** for business operations

### 📋 Related

- Part of comprehensive health check optimization across all FOGIS services
- Addresses root cause of log noise rather than filtering symptoms
- Maintains full monitoring capabilities while reducing operational overhead
- Reference: `HEALTH_CHECK_OPTIMIZATION_SUMMARY.md`

### 🔍 Code Review Focus

- Verify health check functionality preserved
- Confirm business operation logging unaffected
- Validate middleware exclusion logic
- Check performance timing accuracy

---
Pull Request opened by [Augment Code](https://www.augmentcode.com/) with guidance from the PR author